requirements.txt
pdfplumber
pypdf2
python-dotenv
pytest
streamlit
